price data type change to flaot from double or int

+++ src/main/java/sfa/product_service/controller/ProductController.java
@@ -29,7 +29,7 @@ public class ProductController {
     }
 
     @GetMapping("/getByIdAndPriceType/{id}")
-    public ResponseEntity<Double> getProductByIdAndPriceType(@PathVariable Long id, @RequestParam String priceType) {
+    public ResponseEntity<Float> getProductByIdAndPriceType(@PathVariable Long id, @RequestParam String priceType) {
         return new ResponseEntity<>(productService.getProductPriceByIdAndType(id, priceType), HttpStatus.OK);
     }
